move error-page to folder pages
